refactor(calendar): EventDetailModal sous 300 lignes — éclatement complet (H8)
- composable useEventDetail (computeds type/titre/statut, capacités par rôle, formatage, emitAction)
- sous-composants présentationnels EventSeanceDetails + EventEvaluationDetails + EventDetailActions (CSS verbatim)
- orchestrateur 744→202 lignes
- tests: composable (7) + actions (5) + séance (3) + éval (3)
- dette pré-existante conservée verbatim : console.log de debug dans canJoinVisio
- CSS chrome de detail-row dupliqué verbatim entre les 2 sections (scoped)
